requirements.txt
torch
torchvision
opencv-python
numpy
scikit-learn
Flask
Flask-Login
Flask-SQLAlchemy
Pillow
